Arduino result percent() return uint32
Otherwise the format string doesn't work and I don't feel like changing
the format string. decimals in the percentage are dumb was what past me
thought, and I'm inclined to agree

+++ src/proffieconfig/tools/arduino.h
@@ -36,7 +36,7 @@ namespace Arduino {
 
         int32 used{-1};
         int32 total{-1};
-        inline float64 percent() const { return (static_cast<float64>(used) / total) * 100.0; }
+        inline uint32 percent() const { return (static_cast<float64>(used) / total) * 100.0; }
     };
 
     /**
